requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
branca==0.8.2
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distlib==0.4.0
distro==1.9.0
einops==0.8.2
filelock==3.24.3
filetype==1.2.0
Flask==3.1.3
folium==0.20.0
frozenlist==1.8.0
fsspec==2026.2.0
geographiclib==2.1
geopy==2.4.1
google-auth==2.48.0
google-genai==1.64.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
identify==2.6.16
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
nodeenv==1.10.0
numpy==2.0.2
opencv-python-headless==4.11.0.86
packaging==26.0
pillow==10.4.0
platformdirs==4.9.2
pre_commit==4.5.1
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pypdfium2==4.30.0
python-discovery==1.1.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
setuptools==82.0.0
sniffio==1.3.1
surya-ocr==0.17.1
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==21.1.0
websockets==15.0.1
Werkzeug==3.1.6
xyzservices==2025.11.0
yarl==1.22.0
